+++ app/assets/javascripts/pairer/application.js
@@ -1,5 +1,4 @@
 //= require rails-ujs
-//= require bootstrap-sprockets
 
 window.init = function(){
   $('form').attr('autocomplete', 'off');